使用封装的RedisUtils

+++ marathon-common/src/main/java/com/newera/marathon/common/constant/RedisConstant.java
@@ -2,5 +2,5 @@ package com.newera.marathon.common.constant;
 
 public class RedisConstant {
     //captcha有效时间
-    public static final int CAPTCHA_EXPIRY_SECOND = 60 * 5;
+    public static final Long CAPTCHA_EXPIRY_SECOND = 60 * 5L;
 }
